+++ packages/dso-toolkit/src/components/tree-view/index.ts
@@ -1,0 +1,4 @@
+export * from "./tree-view.args.js";
+export * from "./tree-view.models.js";
+export * from "./tree-view.stories-of.js";
+export * from "./tree-view.demo.js";
